requirements.txt
google-api-python-client
pandas
nltk
scikit-learn
matplotlib
joblib
flask
flask-cors
requests
python-dotenv
waitress
textblob
chat-downloader

